+++ project_config/webpack/loaders/extract_css_loader.js
@@ -3,14 +3,17 @@
 
 var ExtractTextPlugin = require('extract-text-webpack-plugin');
 
-module.exports = () => {
+module.exports = config => {
     return {
         test: /\.css$/,
         use: ExtractTextPlugin.extract({
             fallback: 'style-loader',
             use: [
                 {
-                    loader: 'css-loader'
+                    loader: 'css-loader',
+                    options: {
+                        modules: !!config.enableCssModules
+                    }
                 }
             ]
         })